Move benchmark job to GHA (#3830)
Summary:
This PR adds a new workflow for benchmarking. Due to security reasons the process of uploading (requires secrets) the results is in a separate workflow that does not execute user controlled code.

This workflow has to be on the default branch to run so it is not possible to test it here but I have tested it on my fork. The results are visible here: https://velox-conbench.voltrondata.run/

I have added a step to remove any benchmark binaries which are identical between baseline and target to avoid running the same benchmark twice,

For the same reasons I have reduced the workflow on merge to main to only run baseline, previously it was building and running the same benchmarks twice. It will run all benchmarks on push to main.

This is functional but leaves out the comparison of results and re-runs, getting the data collection back online was my main priority for now.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3830

Reviewed By: pedroerp

Differential Revision: D42853790

Pulled By: kgpai

fbshipit-source-id: 67ff8f6d56db07807b3338de5a3fc920c53037d0