Don't rely on workflow API for pr number (#3883)
Summary:
Previously I was using the workflow API and the `pull_requests` key to get the pr number. It  seems that this key is not always populated (empty for fork prs?).

This pr adds a workaround by exporting the pr number as a file.

Tested on my [PR](https://github.com/assignUser/velox/actions/runs/4055757952/jobs/6979324722).

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3883

Reviewed By: Yuhta

Differential Revision: D42890349

Pulled By: kgpai

fbshipit-source-id: 02c3bb2a4a2800065d7c87b6570246626ef29400