Add support for nested expression re-use to the fuzzer (#3669)
Summary:
This adds support for re-using expression in the fuzzer that have
a nesting level greater than 1. It does so by adding a utility
class that can be used to keep track of all generated expression
and a provides an API that allows for randomly picking an expression
that has a specific return type and a nesting level less than or
equal to a specified limit. It ensures that all expressions that
are valid candidates have an equal probability of selection.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3669

Test Plan:
Added unit tests for the new utility class.
Also verified manually that multi-level expression are being re-used.

Reviewed By: kagamiori

Differential Revision: D42434715

Pulled By: bikramSingh91

fbshipit-source-id: 2883a291efa0b7917b9d9774752f76cf0b825c1e