Add array_frequency to skip function list of expression fuzzer (#3911)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/3911

The array_frequency() function is implemented with folly::F14FastMap that doesn't guarantee deterministic ordering of elements. This leads to mismatch of the fuzzer test results that have the same elements but with different orders. So we add array_frequency() to the skip function list for now.

This fixes https://github.com/facebookincubator/velox/issues/3906.

Reviewed By: bikramSingh91

Differential Revision: D42944822

fbshipit-source-id: 50a574a8cb4bb1f70e9251c61405d3b5065528e6