Fix boundary check in MmapAllocator::SizeClass::isInRange (#3893)
Summary:
MmapAllocator::SizeClass::isInRange checks that a given pointer falls on a page boundary.

```
    if ((ptr - address_) % unitSize_ != 0) {
      VELOX_FAIL("Pointer is in a SizeClass but not at page boundary");
    }
```

Here, ptr - address_ is the number of bytes while unitSize_ is the number of
pages. Hence, the check is incorrect as it needs to convert number of pages to
number of bytes before computing the mod.

fix https://github.com/facebookincubator/velox/issues/3892

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3893

Reviewed By: mbasmanova

Differential Revision: D42932962

Pulled By: xiaoxmeng

fbshipit-source-id: 0df9016f543fa639bba66aad44370e5e3023cf4e