Consolidate common sub-expression evaluation logic (#3908)
Summary:
CSE evaluation logic in Expr.cpp was split between
checkUpdateSharedSubexprValues and checkGetSharedSubexprValues method an
interleaved with regular evaluation making it hard to following. This change
consolidates the logic in a single evaluateSharedSubexpr method.

More refactoring is needed to make this logic easier to read and test. This
is a first step.

Depends on https://github.com/facebookincubator/velox/issues/3900

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3908

Reviewed By: xiaoxmeng

Differential Revision: D42936426

Pulled By: mbasmanova

fbshipit-source-id: f86cb50bd5276f01c6d4e93cc58e980e8e45babe