Add support for emit (output ordering) (#3543)
Summary:
See https://substrait.io/relations/basics/#emit-output-ordering for the definition of "emit (output ordering)" in Substrait.

This PR includes the initial integration of emit feature to aggregate, filter and project operations. This work may have follow up work to integrate non-roundtrip test cases to evaluate Substrait to Velox path more thoroughly. The reason for a follow up is we expect to introduce a simplified readable Substrait plan for better usability. Since this is still work in progress we expect to make a follow PR in the future. With that approach we can lessen the text and  add more test cases.

Also existing test cases were updated with the inclusion of projection correction and emit inclusion.  The updates were not done in the code level, but in plan level. The integrated emit feature is added to plans to retain the validated intentions of the test cases. Also we have included validation for emit usage with read operations. We have restricted users in using emit with read nodes since we expect the users should select the required fields while defining the input schema.

## PR Includes

In Substrait, the projection would append fields to input fields [0], in the order such that input fields then the projected fields. The current Velox/Substrait integration doesn't handle it [1]. So the first change introduced in this PR is correcting that. The correction is as follows.

The version in the main branch doesn't contain the Substrait projection spec accurately. It doesn't take the fields from the input to the projection node. This PR corrects that. This is the first change.

The second change made in this PR is integrating the `Emit` [2] feature in Substrait to the Velox-Substrait consumer. The idea is to decide what are the outputs you expect from a particular relation. This is an attribute in `RelCommon` [3] which means each relation has this.

This PR doesn't introduce new test cases but uses the existing test cases to validate the change. But we introduce the emit attributes to the existing plans to validate the intended objective of test cases due to the project correction introduced in this PR.
This was the motivation to include the emit feature within the same PR.

## Note

The project definition update on this PR could be a breaking change to other projects using the current definition on the upstream. To avoid breaking test cases or changes to applications, the easiest change would be to introduce emit to project nodes just to output the projected fields. But, it is recommended to adopt the project definition in Substrait when working with Substrait plans.

## Example

Let's assume the following relational algebra operators are part of a particular query.

Schema -> `schema({ ('a', i32), ('b', i32) })`

`read` -> `project({(a+b), (a-b)})` -> `filter(field(3) > 3)` # (implying filter needs to be applied to `a-b` output)

Substrait output Schema after each relation;

Read -> `schema({ ('a', i32), ('b', i32) })`

Project -> `schema({ ('a', i32), ('b', i32), ('a+b', i32), ('a-b', i32) })`

Filter -> `schema({ ('a', i32), ('b', i32), ('a+b', i32), ('a-b', i32) })` # (would be doing the comparison on `a-b` column)

This is the expected schema change in Substrait.

If just the projected expressions are required, it should be defined as follows;

`read` -> `project({(a+b), (a-b)}).emit([2, 3])` -> `filter(field(1) > 3)` #  In Projection, 0 -> `a`, 1 -> `b`, 2 -> `a+b`, 3 -> `a-b`. Note that you may want to update the filter field accordingly.

After including emit to projection, the output schema for each relation would be;

Read -> `schema({ ('a', i32), ('b', i32) })`

Project -> `schema({('a+b', i32), ('a-b', i32) })`

Filter -> `schema({('a+b', i32), ('a-b', i32) })`

[0]. https://substrait.io/relations/logical_relations/#project-operation
[1]. https://github.com/facebookincubator/velox/blob/main/velox/substrait/SubstraitToVeloxPlan.cpp#L155-L169
[2]. https://substrait.io/relations/basics/#emit-output-ordering
[3]. https://github.com/substrait-io/substrait/blob/main/proto/substrait/algebra.proto#L16-L22

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3543

Reviewed By: Yuhta

Differential Revision: D42963621

Pulled By: mbasmanova

fbshipit-source-id: 2ca84fef2ddc501c29758260f1d64ab474b8ed1b