Choose merge-base from upstream (#4007)
Summary:
Previously we required the fork main to be in-sync with the upstream main, this assumption is not valid. This PR chooses the merge-base based on the upstream main which fixes the issue of an outdated merge base. I also added an error message to make it clear when finding the merge base failed.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4007

Reviewed By: Yuhta

Differential Revision: D43161296

Pulled By: kgpai

fbshipit-source-id: 25ff2ff714c89014137dfa0b8e8111416a372444