Fix DuckDB parse for BOOLEAN literal (#4023)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4023

DuckDB parser is returning cast expression (e.g. `cast('t' as BOOLEAN)`) for BOOLEAN literal.  Fix this by restoring it back to constant expression.

There is a caveat with the fix, that we will automatically constant-fold `cast('t' as BOOLEAN)` into `TRUE` constant expression in the parser.  This should not affect functionality, but we may notice a discrepancy in the expression string we input vs what the `ExprSet` eventually gets.

Reviewed By: mbasmanova

Differential Revision: D43194736

fbshipit-source-id: 2c69e5fc9802129d7583173fa3c7d5acfea58954