Add Optimizer for validating and optimizing
The LogicalPlanner performs various functions including creating,
validating, and optimizing plan nodes. The creation of plan nodes
are dependent on the analyzer artifacts, while validation and
optimization are core engine responsibilities.
To separate these responsibilities, we are introducing an optimizer
class to manage the validation and optimization of plan nodes.

This change also adds a runtime stats optimizerTime for measuring
time taken to validate and optimize plannodes.