Fix race condition in InternalResourceManager initialization
Throw an exception if a query tries to create resource groups before the
configuration manager has been loaded.  This fixes a race condition
where a query can start constructing resource groups before the
ResourceGroupConfigurationManager is loaded.  This would cause some
resource groups to get created in the legacy manager and never recreated
when the new resource manager is loaded. This is one theory about why we
are seeing some clusters get into a state on restart where all queries
fail with a NullPointerException in
ReloadingResourceGroupConfigurationManager.