Fix expression compilation and common sub-expression evaluation (#4057)
Summary:
Currently a bug exists where an expression when marked as a
common sub-expression(CSE) does not have metadata recomputed on it.
This re-computation is important since part of
metadata(distinctFields) is dependent on whether an expression is a
CSE. This happens when an expression is used twice and the second
instance of it is a standalone expression tree itself.
Eg, "concat('foo_',upper(c0))", "upper(c0)". The absence of distinct
fields can result in peeling not being performed every time before
CSE optimization is employed and this can result in all sorts of
issues like wrong results or hitting exceptions. This patch attempts
to fix this bug.

Additionally, a change is added to skip re-computing metadata for
children of an expression if they are a CSE as after this change
its always guaranteed that their metadata will be current so
this will avoid doing redundant work.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4057

Test Plan: Added a unit test

Reviewed By: mbasmanova

Differential Revision: D43367664

Pulled By: bikramSingh91

fbshipit-source-id: d533cb8aed6379630685b7ab96b60241b1e05fac