Remove usage of variant::inferType from variantArrayToVector function (#4059)
Summary:
variant::inferType() is error-prone as it infers logical type (instance of Type
class) from physical type (TypeKind enum), which is not always possible.

I'm going to create a series of PRs to remove usage of variant::inferType()
and then remove the method itself. This is first PR in that series.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4059

Reviewed By: xiaoxmeng

Differential Revision: D43370448

Pulled By: mbasmanova

fbshipit-source-id: 6344fcebb34061e769c0f7565d670cc8408c4153