Fix copy exception when wrappedVector.size < rows.size . (#4124)
Summary:
Currently we will fault , when the wrappedResult has a smaller size than rows and we need to copy the results to a flat results Vector. This situation typically happens when the peeledResult has a lot of trailing nulls and thus applyRows ends up lesser in size than rows. In most such cases its unlikely resultVector is present and is flat which is probably why this bug wasnt caught sooner.
This fix ensures we only copy for applyRows , since we ensure that we add trailing nulls in evalAllImpl.

Fixes https://github.com/facebookincubator/velox/issues/4077 .

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4124

Reviewed By: Yuhta, kagamiori

Differential Revision: D43558528

Pulled By: kgpai

fbshipit-source-id: 590b629a5baa27f7a3936a68488a5093e7570544