Optimize TopN by skipping rows equal to the current top row (#4135)
Summary:
When an incoming row equals the top row in the TopN heap, the incoming row can
be skipped rather than pushed to the heap.

The unnecessary pop and push operations decay performance significantly if there
are lots of duplicate rows.

For example, a test that returns top 10 rows from 5000 rows of zeros runs in
1110 ms before optimization and in 799 ms after optimization.

The performance improvements will be more significant if the number of rows is
extremely large.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4135

Reviewed By: Yuhta

Differential Revision: D43688224

Pulled By: mbasmanova

fbshipit-source-id: 4b67214da60926413346ba4f0fd2beb612b2c49a