Remove check for null map keys in cast-json-to-map function (#4159)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4159

The check for map keys not being null is unnecessary for the cast-json-to-map
function because an input Json-map that contains non-string keys (e.g., "{null:1}"
or "{1:2}") are rejected by folly::parseJson() directly before reaching the casting
code. This diff removes the unnecessary check and move the corresponding test
cases to JsonCastTest.toMap.

Reviewed By: mbasmanova

Differential Revision: D43680360

fbshipit-source-id: 18be65317ea29555a17a348b520fdb537cf67d51