Refactor Thrift Transport for Parquet Metadata Access (#4160)
Summary:
This PR refactors the Thrift Transport for Parquet Metadata access. It uses `ThriftTransport` as an interface and introduces `ThriftBufferedTransport` and `ThriftStreamingTransport`.

`ThriftStreamingTransport` takes in a `SeekableInputStream` as input for Thrift parsing. This can be used for Parquet Page Header parsing. This optimization is able to reduce the deep copy in `readPageHeader()`. And it is also the prerequisite to fix the incorrect page header length issue.

`ThriftBufferedTransport` takes in a consecutive memory space as input for Thrift parsing. This can be used for Parquet Footer parsing, because the footer is at the bottom of the file and we need to also take care of `footer length` and `PAR1`, from the bottom to top.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4160

Reviewed By: kgpai

Differential Revision: D43772723

Pulled By: Yuhta

fbshipit-source-id: f795149b28332effbc87ed05f237d50f5b5ac7be