Restrict signatures of Cast in expression fuzzer (#4162)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4162

Instead of rejecting expressions of unsupported types during signature binding,
CastExpr attempts to evaluate the unsupported expressions and throw during the
processing. This exception cannot be masked by Try expression. On the other
hand, expressiion fuzzer assumes all generated expressions are supported and
hence asserts that retry-with-try should never throw. So this diff restricts the
signatures of Cast in expression fuzzer to a subset of supported ones for basic
types. In addition, the statistics of the usage of Cast in fuzzer was not correct.
This diff also fixes this issue. In a subsequent diff, we'll make CastTypedExpr
maintain a set of supported signatures and expose them to expression fuzzer.

This diff fixes https://github.com/facebookincubator/velox/issues/4145.

Reviewed By: mbasmanova

Differential Revision: D43686197

fbshipit-source-id: 6ed7dfa912107d2cd7210866e1de067854d00812