Read Parquet Page Header with ThriftStreamingTransport to Fix the Incorrect Header Length (#4108)
Summary:
The size of Parquet Page Header remains unknown until the Header is parsed because it is in Thrift format. Without this PR, the size is estimated by `sizeof(PageHeader)` but the actual Parquet Page Header might be larger. It can be reproduced by using the given `largePageHeader`. This PR takes the advantage of `ThriftStreamingTransport` with the input of `inputStream_` to avoid estimating the Parquet Page Length. In this case, the incorrect Parquet Page Header length can be resolved.

Also, this approach avoids unnecessary deep copy when reading Parquet Page Headers. Without this PR, deep copying might happen when preparing the input data from in-consecutive chunks  for `ThriftBufferedTransport`.

This PR also introduced the related Parquet page header unit tests on small, large, and corrupted Parquet Page Headers.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4108

Reviewed By: pedroerp

Differential Revision: D43855209

Pulled By: Yuhta

fbshipit-source-id: 242de9f9b80ae9cdd5c27a99c9640893f3329df9