Fix null ptr issue for empty join optimization
Currently we check whether hash table is empty when calling needsInput
for LookupJoinOperatop, and set finishing to true if it's empty.
The hashtable may be released after finishing is set to true. And this
can be a problem if needsInput is called again, which will try to access
the hashtable and leads to null pointer exception.
This PR fix it by guarding the check with finishing to be false.