Fix thread safety in prepare of ConnectorQueryCtx (#4233)
Summary:
DataSources may be prepared asynchronously for one TableScan on multiple threads. This may entail preparing a remaining filter ExprSet. This is done with a ConnectorExpressionEvaluator. This in turn uses a single ExecCtx for expression preparation, which is not thread safe. Furthermore, the DataSources being asynchronously prepared are in a queue shared between Drivers. The DataSource could end up with a Driver other than the one that created it.

At read time, each async prepared DataSource is merged into the DataSource f the TableScan doing the reading.

The remaining expression is prepared for each DataSource. This is needed in order to know what columns it requires. Also, metadata level pruning could use expressions.

Therefore, we document that ConnectorQueryCtx and its component ConnectorExpressionEvaluator are not thread safe. We create a ConnectorQueryCtx for each async prepare of a DataSource. We make ConnectorQueryCtx independent of the creating OperatorCtx. In specific, it does not share its ExecCtx, which is strictly single-threaded. It does use the pool of the creating OperatorCtx, which is safe since the Pools are destructed only on task destruction. The expression evaluator is further owned by the ConnectorQueryCtx to make the whole decoupled from the creating context.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4233

Reviewed By: xiaoxmeng

Differential Revision: D43919297

Pulled By: oerling

fbshipit-source-id: 042cd5cd32dfb0d065b040577d640b7a80f65f63