Rewrite inline constructor, delete conversions which can dangle (#3505)
Summary:
`StringView` has a constructor which accepts just the size. This constructor is only used once in the codebase to construct an inline string view into which characters can be directly written. I've rewritten this to an explicit factory function which encapsulates the `const_cast`.

Some of the conversion operators and accessors could lead to dangling pointers in the presence of inlined views. For example, the expression `StringView("").data()` compiles but results in a pointer into the newly destroyed temporary. Any accessor or conversion which may return a pointer into the view itself should be deleted for rvalue references to prevent this.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3505

Reviewed By: tjsmith-meta, kgpai

Differential Revision: D43883019

Pulled By: pedroerp

fbshipit-source-id: e24f107964fb39dd7258d350bc9635943c0d987a