Fix peeling of constant vector smaller than selected inner rows size (#4277)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4277

Fuzzer found a bug where two input vectors of an expression are a dictionary vector and a
constant vector whose size is smaller than the size of selected base rows of the dictionary
one. The current code peels off common encodings (i.e., the dictionary layer) and save the
dictionary base vector together with the original constant vector to EvalCtx for later access.
The subsequent evaluation is performed on the selected inner rows of the dictionary vector,
which is larger than the constant vector size and hence causes out-of-bound access.

This diff fixes this issue by resizing the constant vector to the size of the selected inner rows
after peeling if the constant vector size is smaller. This resizing happens before the constant
vector is saved to EvalCtx.

This diff fixes https://github.com/facebookincubator/velox/issues/4262.

Reviewed By: mbasmanova

Differential Revision: D44059058

fbshipit-source-id: f618716303f1ec8f7da719f4394cd3c4801e0c28