Fix filter evaluation for CHAR(n) in SliceDictionarySelectiveReader
When the filter for a column of CHAR(n) type is passed to the stream
readers, the padded spaces would be removed. Therefore when testing
a value's length while evaluating the filter, it's incorrect to use n.
Instead, the filter's length shall be compared with the value's real
length, which was already calculated and stored in dictionaryOffsetVector.

Fixes https://github.com/prestodb/presto/issues/19138