Improve Null Propagation Precision (#4269)
Summary:
Many expressions were not recognized as null propagating, thus reducing effectiveness of running on distinct values in dictionaries. With this change, a a function call to a default null behavior function is null propagating if the set of columns null propagating arguments depend on covers the set of arguments non-null propagating arguments depend on.

For example substr(c, 1,, if (length(c) > 2, length(c) - 1, 0)) is null propagating but was not recognized as such because the if is not null propagating. It depends however on the same set of columns 'c' depends on, hence a null in c makes a null in substr.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4269

Reviewed By: Yuhta

Differential Revision: D44026162

Pulled By: oerling

fbshipit-source-id: 70ce7be0e9e88e67fd9a7b447d6e8b1b67822c68