Allow passing context to ISerializable::deserialize (#4307)
Summary:
A follow-up PR adds serialization for expressions. To serialize
ConstantTypeExpr, we use VectorSaver to serialize the constant vector. To
deserialize, we need to reconstruct the constant vector, which requires a
memory pool. Hence, we need a way to pass memory pool to
ISerializable::deserialize.

Part of https://github.com/facebookincubator/velox/issues/4303

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4307

Reviewed By: xiaoxmeng, pedroerp

Differential Revision: D44119995

Pulled By: mbasmanova

fbshipit-source-id: f6e0c9e7b9efcd97e2b25a1beb0f9a298c7efc31