Add benchmark for wrapping and peeling with Exprs (#4196)
Summary:
Runs a sequence of 4 filter + project pairs on inputs with 4 bigints or 4 strings. Each project asses through 3 values of the 4 and modifies one with an expression.

String values are run with flat and dictionary encoding to allow evaluating on distinct values and memoizing.

See comments in FilterProjectBenchmark.cpp.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4196

Reviewed By: Yuhta

Differential Revision: D43849889

Pulled By: oerling

fbshipit-source-id: aab5506245701c524643b15b517f7806fa54dcce