Fix: make vector copy functions resilient to invalid offsets when count is 0. (#4393)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4393

it's a common practice to call copy with count 0, for example when copying
arrays and maps. i.e:
```
auto inputOffset = inputMap->offsetAt(index);
 auto inputSize = inputMap->sizeAt(index);
  combinedValues->copy(
            inputMap->mapValues().get(), offset, inputOffset, inputSize);
```
In the case above inputOffset could be anything if inputSize is 0.
This makes some checks fail inside the copy functions.
This diff makes the copy resilient to invalid offset and inputOffset if inputSize ==0
by early exiting. The goal is to reduce such bug occurrences by providing a more resilient API.

Reviewed By: kagamiori

Differential Revision: D44317822

fbshipit-source-id: 27fa9b4903d1d6487704f6610c6c76e7ceca6250