Quantize exchange allocation size
Quantizes allocation sizes of IOBufs for PrestoExchangeClient. The
minimum size is 4K and the maximum is 64K. The allocations for
exchange IOBufs will be chains of 64K chunks or larger if individual
chunks produced by Proxygen are larger.

The benefit is that allocation is tracked by MemoryPool and for mmap
backed memory will always come from MmapAllocator and not malloc. This
means no fragmentation and no rise in resident set size. Cache will be
evicted to make space for exchange and RSS will not go up because of
alternating demand between malloc and MmapAllocator.

Also, the frequency of allocation drops for receiving long messages
which saves some CPU.