Add snapshot to presto-verifier
Verifier runs Control and Test queries and matches the results for
verification.
This PR improves the Verifier and allows it run in query-bank mode,
that is, save the result in database as a snapshot, and subsequent
runs only need to re-run the Test queries and match the Test result
with the saved snapshot.