Fix classloading issue with OutputNode plannode
OutputNode plan class is used for creating logical plan. Now logical
plan creation is part of the pluggable analyzer interface, respective
analyzer plugins are expected to use OutputNode class.

Previously, the OutputNode class was extended from InternalPlanNode,
which required it to be available with AppClassLoader. However, since
the analyzer plugins now instantiate the OutputNode class, it and its
dependencies are loaded by PluginClassLoader, making it unavailable
for the rest of the AppClassLoader.

To fix this issue, we are extending OutputNode from PlanNode instead
of InternalPlanNode and also moving it to the spi package, as it is
anyways expected to be leveraged by plugins. This also whitelists
this class for PluginClassLoader.

The other alternative would have been to add
com.facebook.presto.sql.planner package to the whitelists for
PluginClassLoader, however that would have added to many classes to
the whitelist which is not necessary.