Fix error semantics in null-propagating switch (#4375)
Summary:
A null propagating switch can have inconsistent behavior with nulls and errors. If a condition on a row with nulls has an error, we do not see the error if the arguments are known. If they are lazy, we can hit the error before loading the lazy. Switches do not generally preload lazies because the whole point is to avoid loading lazies that are never accessed.

A null propagating switch however requires that all thens and else depend on the same columns, so load all lazies regardless of which way the condition goes. Therefore for null propagating switches, we first load the lazies, then resolve the null rows and then run the switch for the remaining rows.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4375

Reviewed By: kagamiori

Differential Revision: D44281206

Pulled By: oerling

fbshipit-source-id: 1875a3e58899099c6d3922b0ad0575c5ba3eb0f5