Fix error vector handling in Coalesce (#4394)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4394

Expression fuzzer found that the result of an expression `try(coalesce(foo(...), or(...)))` is incorrect in that rows throwing in foo() have non-null values in the result. The root cause is that Coalesce passes the error vector from foo() to or(), whereas or() clears the error vector at rows that are not among the initial active rows or not among the final active rows. This diff fixes the problem by deselecting error rows after evaluating every argument of Coalesce and making Conjunct preserve pre-existing errors outside of rows that it evaluates on.

This diff is the first fix for https://github.com/facebookincubator/velox/issues/4337.

Reviewed By: mbasmanova

Differential Revision: D44321132

fbshipit-source-id: d1ec2f254b36773bbe4a20aaff51005f79a4cae3