Add support for merge join without grouped execution
Previously, we could only plan a merge join when a stage ran with
grouped execution. This was due to sorted bucketed tables not being
processed in an ordered way without grouped execution. To resolve this
we add a merging exchange if the input is not single stream.  We also
add an optimizer that runs on the plan fragment after grouped execution
is determined to remove the merging exchange if the stage is running
with grouped execution.