Optimize null-aware anti join with filter (#4433)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4433

In current implementation, we still try to apply filter on build side rows with null keys even when we know from `HashBuild` that there is no null key on build side.  This change optimizes this step away, and brings the performance of null-aware and regular anti join to be the same on null-free data.

We also optimize the traversal of all null key rows in row container.  Instead of going through all the rows in container and filter them by selectivity later, we reuse the linked list for duplicated key values, and traverse this linked list only to get all null key rows efficiently.

Also fix some minor issues in TPC-H benchmark.

Reviewed By: oerling

Differential Revision: D44388495

fbshipit-source-id: d8ec3ba09e0f4bd479267574df46a52112c1a4bb