Add test for native execution of functions registered via a json file
The support for this feature was added via
`JsonFileBasedFunctionNamespaceManager` in #18630.

The introduced test fires a query `SELECT json.test_schema.eq(...) ...`
and asserts successful registration of the external function in presto
function register as well as successful registration+execution of the
function in presto native(velox). On native side, the function is
conditionally registered via the introduced config flag
`register-test-functions` that's only true for testing.