Fix: make CSE and peeling orthogonal. (#4441)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4441

CSE implementation requires that when evaluateSharedSubexpr on the
same expression the inputs are the same to the expressions.

This can break when peeling happens, for example for the
expressions f(b(c0), c0) and b(c0), it is possible that b(c0) is evaluated on the
 peeled c0 the first time it is executed BUT  on the non-peeled c0 the second
time (see the tests for example). Furthermore, this property is not trivial to
maintain.

To solve this problem, I changed CSE to capture the inputs and the rows during
the expression evaluation and then look up shared results using both of them
instead of just the rows.

This does not prohibit vector reuse, because the inputs vectors anyway are not
 re-usable since they are assumed to be held by the expression evaluation
caller.

This solves https://github.com/facebookincubator/velox/issues/4351

Reviewed By: Yuhta

Differential Revision: D44442738

fbshipit-source-id: 32646db947625dbd9d461bb0d84df9beec328f79