Fix document generation job (#4569)
Summary:
Fixes https://github.com/facebookincubator/velox/issues/4568

Job failed at the end because the condition that executes a commit and
push only checks for a non-trivial git status to check whether any
changes need to be added. However, it fails to account for untracked
files which also show up in the output. This results in the git commit
command to fail with an exit code of 1 and eventually fail the job.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4569

Reviewed By: laithsakka

Differential Revision: D44849332

Pulled By: bikramSingh91

fbshipit-source-id: a0eeb666e33373c2b00c8be2414552abea4d1698