Fix error vector handling in SwitchExpr (#4580)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4580

Expression fuzzer found that the result of an expression try(switch(foo(...), or(...), ...)) is incorrect in that rows throwing in foo() have non-null values in the result. The root cause is that SwitchExpr passes the error vector from foo() to or(), whereas or() clears the error vector at rows that are not among the initial active rows or not among the final active rows. This diff fixes the problem by deselecting error rows after evaluating every argument of SwitchExpr.

Checked that the behavior after the fix is the same as Presto's.
```
SELECT
    try(CASE WHEN 5 / a > 0 THEN 1 WHEN 4 / (a + 1) > 0 THEN 2 ELSE 3 END)
FROM (
    VALUES
        (1),
        (0)
) AS t(a)
```

| _col0 |
| 1 |
| NULL |

This diff is the second fix for https://github.com/facebookincubator/velox/issues/4337.

Reviewed By: bikramSingh91

Differential Revision: D44851673

fbshipit-source-id: 92b265ffefd2dedcd13d3af7037fabef9fc54774