[native] Set string buffer size in PartitionAndSerialize operator
PartitionAndSerialize didn't set the size for the string buffers and produced
invalid string vectors. This bug is similar to the one fixed in
https://github.com/facebookincubator/velox/pull/3265. This bug may exist in
TableScan as well: https://github.com/facebookincubator/velox/issues/4608

There is no convenient way to add a test for this fix. We should add a method to
Velox to check integrity of the vector and use it in TaskCursor to ensure all
vectors produced by query plans used in tests are valid.