Fix TableScan setBatchSize (#4463)
Summary:
Right now, when the estimated row size is in the range of [1K,100K)
bytes, the TableScan operator outputs batches of 100 rows. This could
create many small output batches. This change fixes it and caps
the batches by size 10MB instead of the number of rows.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4463

Reviewed By: xiaoxmeng, Yuhta

Differential Revision: D44532427

Pulled By: gggrace14

fbshipit-source-id: 099b245c9c7e76df464d898e8d92d8392ae781db