Fix: avoid out of bound access in utf8proc_codepoint. (#4626)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4626

when utf8proc_codepoint is called with an invalid string where the last x
bytes are a prefix of an incomplete utf8 char, then the function
utf8proc_codepoint accesses memory outside the input string.

This diff changes utf8proc_codepoint to take the `end` of the input string
as input and return invalid codepoint for incomplete characters described
above.

this fixes.
https://github.com/facebookincubator/velox/issues/4554

Reviewed By: kgpai

Differential Revision: D45026804

fbshipit-source-id: e52ed853352d38f0d9973169630a42fdcb26d76e