Fix CREATE/DROP TEMPORARY FUNCTION
This was broken by https://github.com/prestodb/presto/pull/16699 and has
not worked since release 0.268.  The CreateFunctionTask and
DropFunctionTask should be adding functions directly to the
QueryStateMachine rather than keeping a special map in the task. This
commit basically reverts the changes that #16699 introduced for create
and drop function.