Clear HasFilter status in AddSplit (#4597)
Summary:
When dynamic filter on partition key is pushed down to table scan, sometimes the result becomes empty.

The root cause is we add the partition key filter in `HiveDataSource::addDynamicFilter` first, then call `ScanSpec::resetCachedValues`, and in `ScanSpec::reorder`, we decide the partition key filter is a real filter and set hasFilter_ to true, because the constant value is only set later in HiveDataSource::addSplit.

So we have to reset hasFilter_ for ScanSpec after set constant value.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4597

Reviewed By: kgpai

Differential Revision: D45085314

Pulled By: Yuhta

fbshipit-source-id: 7be07833eb05ceea183e0378b670663c946e5393