Use estimateFlatSize for dictionary input vector for LocalExchangeQueue (#4647)
Summary:
LocalPartition is blocked if the total buffered input size of the LocalExchangeQueue hits the max_local_exchange_buffer_size. Since the input is dictionary vector, we should use estimateFlatSize() instead of retainedSize() for size estimate for the input. retainedSize() is the size of the wrapped base vector, and thus is ~15 times larger than the actual size. This makes LocalPartition blocked for every input vector.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4647

Reviewed By: xiaoxmeng, oerling

Differential Revision: D45077999

Pulled By: gggrace14

fbshipit-source-id: 9af42b11cd87438f3b4ca9de478ba3cc7dbd8f99