Fix exception when pushing down TRY expression
Previous commit fixed a bug in isTryFunction. Previously, it would always result to false, because it was comparing QualifiedObjectName vs String. After fixing, we saw some crashes in PredicatePushdown. As previously, all queries were running with it false, I just removed the code which checked that. Added a test for it that we don't pushdown expressions with TRY()