Replace find_library with find_package (#4689)
Summary:
This PR  replaces calls to `find_library` with `find_package`, this enables us to use targets instead of variables for the affected dependencies.

libevent and double-conversion are folly dependencies that we don't need to call `find_library` on as folly handles that internaly. We can add these back in with resolve_dependency later on.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4689

Reviewed By: kgpai

Differential Revision: D45292478

Pulled By: pedroerp

fbshipit-source-id: f4cb8b12822067302a97d79bbdb20f87ceda3280