Refine StringIdMap::makeId (#4684)
Summary:
This PR does two things.

First, in StringIdMap::makeId, the intension of the while loop is to avoid duplicate id
```
  do {
    entry.id = ++lastId_;
  } while (idToString_.find(entry.id) != idToString_.end());
```
However, when lastId reaches ```kNoId - 1```, the next id will be ```kNoId```, which is incorrect.

Second, in StringIdMap::makeId, the following check is unnecessary
```
    if (++entry->second.numInUse == 1) {
      pinnedSize_ += entry->second.string.size();
    }
```
the logic in StringIdMap ensures that ```numInUse``` will never be zero because when ```StringIdMap::release``` decreases ```numInUse``` to zero, it will remove from ```stringToId_``` and ```idToString_```. So it's better to use ```VELOX_CHECK_LT``` to check instead of an if block.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4684

Reviewed By: Yuhta

Differential Revision: D45286483

Pulled By: xiaoxmeng

fbshipit-source-id: 121eb23971900373decfc77be75dd46f6673235d