Add non-local fs support and parameter tuning to velox_tpch_benchmark (#4522)
Summary:
Makes TpchQueryBuilder.cpp accept a data path where each table is represented with a file listing paths to the data files.  Adds a --test_flags_file to flag to velox_tpch_benchmark. This file has one line for each flag to be included in the parameter comparison.  The line is of the form

flag:value1,value2, ... value-n

The workload will be done for all listed values of flag. If there are multiple lines all combinations of flags will be tried.

The results are presented with run times followed by the flag values used for the run, fastest run first.

Detailed output for each run, fastest first follows.

Adds flags for clearing RAM and SSD caches between runs and for adding a warmup run before the measured run.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4522

Reviewed By: Yuhta

Differential Revision: D45220008

Pulled By: oerling

fbshipit-source-id: da3cad371d5fee00ca93b2c2f2a89ad260cf6ab9