Bypass non-reducing partial aggregation (#4588)
Summary:
If a partial aggregation does not reduce data size even after increasing its memory, we stop doing it. Adds a toIntermediate method to GroupingSet and Aggregates. This translates a vector of raw input into a vector of accumulators, each initialized from the single input value on the row.

If toIntermediate is not defined for one or more of the aggregates, GroupingSet::toIntermediate makes temporary rows in the RowContainer and uses the regular Aggregate methods to make a single-value accumulator.

Fixes https://github.com/facebookincubator/velox/issues/3194

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4588

Reviewed By: gggrace14

Differential Revision: D44936760

Pulled By: oerling

fbshipit-source-id: 48ae11b392fa44082a5d38bc1a9816d5091e78e4