Add constant-encoded vectors to PyVelox (#3819)
Summary:
This diff adds support for constant vector bindings in PyVelox. Currently so complex vectors arent supported yet in PyVelox so we only support the basic value constants. Constant Vector is required for expression eval and also for more full fledged fuzzer support in the future.
This PR also adds bindings for simple and flat vectors on all the primitive types . Support for user defined types and vectors corresponding to them will come in a future PR.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3819

Reviewed By: pedroerp

Differential Revision: D43666079

Pulled By: kgpai

fbshipit-source-id: 2d3624104b81ec21d67c258450ab6cc44f2a39a1