Introduce acquireSharedStringBuffersRecursive logic (#4033)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4033

In the simple function interface we cant avoid copying strings from an a
string argument if its nested in complex structure.

For example: subscript for type Array<Varchar>
or an function that filters an array<Varchar> have to copy the strings.

Adding  acquireSharedStringBuffersRecursive allow us to mitigate that.
acquireSharedStringBuffersRecursive acquires ownership for any string buffer
that appears in BaseVector or any of its children recursively.

This diff also document acquireSharedStringBuffers and add a DCHECK that
 the input vector kind is as expected.

Reviewed By: bikramSingh91

Differential Revision: D42890518

fbshipit-source-id: 97089ed1003e6f5c45dbaf6de36529eb14637bc3