Fix the velox promise warning in hash probe processing (#4802)
Summary:
There is VeloxPromise destruction warning log in HashProbe::noMoreInputInternal.
It is because that for the last probed table, the hash probe operator doesn't need to
wait for the other peer probe operators to finish, and it only needs to call
allPeersFinished to tell which one is the last finished hash probe operator. The last
one needs to continue producing the output but all the other probe operators can
just finish processing without waiting at the barrier.

This PR fixes the warning by passing a null future if the caller of allPeersFinished
doesn't intend to wait for the other peers at the barrier. allPeersFinished won't set
the promise for the calling driver.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4802

Reviewed By: tanjialiang

Differential Revision: D45471399

Pulled By: xiaoxmeng

fbshipit-source-id: 29546accebec4f8d86bd8b6aacaa82109be1a386