Fix PickJoinSides handling of local exchanges
Join build and probe sides have different requirements for the local
distribution they need. However, at runtime local exchanges have already
been added, so when we flip join sides, we also need to adjust local
exchanges on the left and right accordingly.

Extract the code that handles this from RuntimeReorderJoinSides to a
common utility class so we can reuse it for this optimizer.

Without this change, queries will have wrong results. The test for
this is in the commit that adds the reoptimization step for AQE.