[Presto-on-spark native] [localMode] Close executor service on shutdown
If we are in localMode, the executor spawns the Executor Airlift application
(which is long running and holds onto resources) on the same JVM

On query completion, the SparkContext shutdown calls the Driver Airlift
application shutdown, but it has no hook to call Executor Airlift application
shutdown. So we hang forever.

This code, prevents this hanging state by explicity calling the
Executor Airlift application to shutdown