Add new field attemptNumber to TaskId
The TaskId class needs to be enhanced to capture retry semantics.
This is applicable in presto-on-spark and more-so in presto-on-spark
native where the TaskId is used to create a remote task on the CPP
process.

By translation, the new string representation of TaskId will contain
5 parts (1 additional) as below

    [QueryId].[StageId].[StageExecutionId].[TaskId].[AttemptNumber]

The attempt number will start at 0 and upto supported number
of retries.

Example:
20230522_133931_02285_ummfz.1.0.21.0
20230522_133931_02285_ummfz.1.0.21.1
20230522_133931_02285_ummfz.1.0.21.2
20230522_133931_02285_ummfz.1.0.21.3

represent the 4 attempts of task 20230522_133931_02285_ummfz.1.0.21

Without the additional attemptNumber field, 2 attempts
of the same task would have identical TaskId violating
the contract between Presto Java process and Presto CPP
process

For non presto-on-spark use-case (presto classic) this field will be set to 0 by default.