Fix decoding of percentiles in ApproxPercentileAggregate::addIntermediate (#4903)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4903

ApproxPercentileAggregate::addIntermediate() receives an input vector of Row type and SelectivityVector `rows` for this input vector. addIntermediate() currently uses `rows` to decode a child of the base vector of the input vector.

This diff fixes it by translating `rows` to the selected base rows before decoding the child vector.

Reviewed By: Yuhta

Differential Revision: D45751685

fbshipit-source-id: 2061d5a371e8ccc2612675b61170e789f45fcce5