Use rethrow for memory reservation failure in mmap allocator (#4976)
Summary:
Make sure we rethrow the caught exception on mmap allocator
failure. We have seen in Prestissimo that the coordinator gets
std::exception on memory capacity exceeded error. The place changed
in this diff is call site where we have seen this problem in prod.
On Mac, throw and rethrow are different but not sure if the other
compiler or platform handles the same way. Added unit test to
catch this as well.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4976

Test Plan: https://www.internalfb.com/phabricator/paste/view/738622280

Reviewed By: spershin

Differential Revision: D45953851

Pulled By: xiaoxmeng

fbshipit-source-id: 19ec224c77febe8234fc682af734b6b6dab1456a