Don't merge regions for vread (#5010)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5010

For wsVRead we need to pass all regions by separate.

`enqueuedToBufferOffset_` was holding an intermediate value between sort and merge. But now, since we may not merge, se save the intermediate result in `offsetToEnqueued_`.

I also realized that, if we don't merge, we don't need to create another `regions` variable for calling vread. We can just use `regions_`.

Reviewed By: Sullivan-Patrick

Differential Revision: D46012935

fbshipit-source-id: eb2a8ad64065afe0ef2a7e0be7e9342ae3b918bc