[native] Make PrestoTaskId backward compatible
Currently, because of #19649 TaskId is expected to have 5 parts.
The 5th part - attemptNumber - is a new addition. For production
use-cases, there is no issue as Java and CPP process will use newer
code. But for testing use-cases, for presto-on-spark, this backward
incompatibility will cause errors. This PR enhances the parsing logic
so that the CPP code is compatible with older Java code. We should be
able to remove this special handling once both java and cpp have caught
up to latest version in all non-prod systems