Fix hang bug in partial aggregation. (#5050)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5050

The bug was triggered in the following way:
- In HashAggregation::addInput() for a partial aggregation
abandonPartialAggregationEarly() returned true and we set partialFull_ to true.
- At the same time grouping set didn't have any new distinct groups and set
newDistincts_ to false.
- HashAggregation operator will then keep returning null from getOutput()
and false from needsInput() thus putting Driver::runInternal into the
infinite loop and hanging the query.

The fix is to detect such corner case and unblock by resetting partialFull_ flag.

Reviewed By: xiaoxmeng, Yuhta

Differential Revision: D46214290

fbshipit-source-id: 8094763b4bb302aae89dd09ffa4ec63bc2fd87c8