Use VELOX_FAIL for unexpected scenario in Task::next() (#4856)
Summary:
The `VELOX_CHECK_EQ(0, blockedDrivers)` used in the current code must
not be true (the condition for entering the `if` branch is
`if (blockedDrivers > 0)`), so it should be replaced by `VELOX_FAIL`
to improve readability.

This patch also fixes some typos.

No functional change in this patch.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4856

Reviewed By: Yuhta

Differential Revision: D46278338

Pulled By: mbasmanova

fbshipit-source-id: 4350bf3ef193365c767e4a31b0b883202d472bf1