Ensure allPeersFinishedPromises are fulfilled. (#5078)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5078

It is possible for the last peer to throw an exception after it got all the promises and before it fulfilled them.
That would destroy promises before they are fulfilled and will result in an incorrect error message to be registered for the Task.
To fix this, we wrap promises into the folly guard.

Reviewed By: xiaoxmeng

Differential Revision: D46300760

fbshipit-source-id: d25bdc2b236fe836828c2bd90bcfb235497751a8