Fix decimal result type when LHS and RHS have different precision in DecimalBaseFunction (#4942)
Summary:
This PR fixes an issue in DecimalBaseFunction when a query would fail with
```
F20230516 07:23:06.503842 810508 BaseVector.h:115] Check failed: dynamic_cast<const T*>(this) != nullptr
*** Check failure stack trace: ***
    @        0x12498f310  google::LogMessage::Flush()
    @        0x124992ca4  google::LogMessageFatal::~LogMessageFatal()
    @        0x12499007c  google::LogMessageFatal::~LogMessageFatal()
    @        0x106bfd9a8  facebook::velox::BaseVector::asUnchecked<>()
    @        0x106bfd9a8  facebook::velox::BaseVector::asUnchecked<>()
    @        0x106bfd9a8  facebook::velox::BaseVector::asUnchecked<>()
    @        0x106c3f1e0  facebook::velox::functions::(anonymous namespace)::DecimalBaseFunction<>::prepareResults()
    @        0x106c3f1e0  facebook::velox::functions::(anonymous namespace)::DecimalBaseFunction<>::prepareResults()
```

This was due to a missing case when LHS and RHS have different precision and if RHS type was long decimal, the result does not have to be a long decimal necessarily. For example, in case of division, LHS `DECIMAL(12, 2)` / RHS `DECIMAL(19,0)` would produce a short decimal result `DECIMAL(12,2)`.

This fixes https://github.com/prestodb/presto/issues/19599

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4942

Reviewed By: amitkdutta

Differential Revision: D46358183

Pulled By: Yuhta

fbshipit-source-id: b47b07d7788396fdad9c94860c3137a6e8177930