Decouple RowContainer and HashTable from Aggregate (#5122)
Summary:
Tight coupling of RowContainer and HashTable to executable aggregate function
makes it difficult to use these data structures in window operators (Window and
TopNRowNumber). The only information RowContainer needs from an aggregate
function is metadata about how much memory is needed for the accumulator and
whether it has any specific alignment requirements. In addition, RowContainer
needs a way to call aggregate function to release out-of-line memory
allocations.

This change introduces an Accumulator class that specifies memory requirements
of an aggregate function and provides a 'destroy' method to release out-of-line
memory. RowContainer and HashTable are modified to use Accumulator instead
of Aggregate.

A slight wrinkle is that Spiller needs access to executable aggregations to
spill intermediate results. It calls Aggregate::extractAccumulators to convert
results into vectors so these can be serialized to disk. To accommodate this
use case, Accumulator class provides access to Aggregate instance. In the
future, we could generalize this interface to allow for spilling non-Aggregate
accumulators as well (e.g. window states).

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5122

Reviewed By: Yuhta

Differential Revision: D46390802

Pulled By: mbasmanova

fbshipit-source-id: 533add52bb3ef8a0c3412224c1ed4046bf7cc9eb