Add to handle memory arbitration failure by aborting query with largest capacity (#5061)
Summary:
Add to handle memory arbitration failure by aborting the query with
the largest capacity instead of simply failing the requestor. This helps
preventing to fail a random victim query which happens to trigger the
failed the memory arbitration. This is helps the distributed query execution
use case such as prestissimo to fail the same query on all the workers.
Here we assume the query use the most memory on all its shards in the
cluster.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5061

Reviewed By: oerling

Differential Revision: D46425567

Pulled By: xiaoxmeng

fbshipit-source-id: c2636a1fc7c6f0e81e897d8d39e35101e4d2786c