Make HttpClient owned by shared_ptr
if HttpClient is owned by unique_ptr from PrestoExchangeSource, an
error can cause the source to be deleted while there is activity on
the client. The client runs activity in event base threads and
registers callbacks. One crash occurs when Proxygen tries to set a
timeout after 'timer_' from HttpClient has been deleted.

The ResponseHandler hereby takes a shared_ptr to the HttpClient so
that its timer and other resources stay live for te duration of
Proxygen callback chains.