Add fuzzer support for feeding preloaded lazy vectors (#4911)
Summary:
This change adds the ability to randomly pick columns in the input
row vector to wrap in lazy and load before feeding to common
evaluation path. It helps to simulate the case where a lazy vector
is loaded in a previous step before being fed to the evaluator.
(It would help catch issues like the one fixed in https://github.com/facebookincubator/velox/issues/4547)
It re-purposed the mechanism already in place that generates a list
of column indices to wrap in lazy and enhances it to use negative
indices as well which represent columns indices that should be
preloaded. This makes the change backwards compatible with repro
files generated before it and ensures minimal code changes.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4911

Reviewed By: kagamiori

Differential Revision: D45776592

Pulled By: bikramSingh91

fbshipit-source-id: 6a9c4304b75da5f35bfcd99f796d9c43be257501