[native pos][tests] Reduce taskInfoUpdateInterval to speed up tests
Currently, the default value of task.info-update-interval is set to 3sec.
We wait atleast 3 secs to receive an update on the CPP task. This slows down
test execution as most test tasks are completed within few milliseconds.

Updating this to 100ms for tests removs the extra 3 sec overhead per task
and speeds up test execution on CI by ~2x and on laptop by ~3x