Add DynamicRowView and its reader (#5075)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5075

A DynamicRowView can represent a row vector of any width of any types, without having to be typed.
It has two functions:
1. size() : width of the struct.
2. at(int i) : return generic view for the field i.

This simplify and enable reading vectors of rows with unknown
or unlimited types. This is also needed  to support row types
in the GenericWriter copy_from operation.

This can't be used in the simple function interface as top level because our
function signature does not support it but it generics can be casted
to it. It is needed to enable row
support in generic writer copy_from operation will be added.

Reviewed By: kevinwilfong, kagamiori

Differential Revision: D46299159

fbshipit-source-id: 8b86a1dc0d9ccabaded98cf77745b852d1318a59