Ignore empty frames in window rank functions (#5161)
Summary:
Previously NULL values were returned for empty frames in window rank functions. This behavior was consistent with DuckDB, but inconsistent with Presto that ignored empty frames.

DuckDB has fixed the behavior in 0.8.0. And Velox is being migrated to it.

Removing the logic to set NULL values for empty frames in rank window functions.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5161

Reviewed By: laithsakka

Differential Revision: D46490085

Pulled By: mbasmanova

fbshipit-source-id: 19c3903385a8d2b1a599eefb3e02a3b6727e2010