Grow DataBuffer size to twice capacity when calling reserve method (#4854)
Summary:
When we call the velox parquet write API in Gluten, the execution time can be 800s when writing 1GB TPC-DS `store_sales`. And we found that the performance bottleneck is the constant reallocating of the memory pool. The reason is that `DataBuffer `is directly reallocated in `reverse()` method , and `memcpy ` must be called every time, resulting in performance degradation. We adjust the size of each reallocate to 2 times in this PR. The execution time can be 3s with this PR.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4854

Reviewed By: mbasmanova

Differential Revision: D46485773

Pulled By: Yuhta

fbshipit-source-id: e3def61abadf65fb41baf9f07a730e429b83347d