Do not allocate dictionary filter cache if there is no filter (#5208)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5208

Dictionary filter cache occupies huge amount of memory for flatmap column when the number of children is large.  We can reduce the memory as well as CPU usage by skipping allocating and initializing the filter cache when there is no filter on children (which is always the case for flatmap read as map).

Reviewed By: oerling

Differential Revision: D46567619

fbshipit-source-id: ec9a0b1c72b863060ddf5439eefd5a07c2a8bfbe