Add max capacity to memory pool to support arbitration (#5037)
Summary:
Add to support max capacity in memory pool which is the hard
capacity limit that a memory pool can grow under memory arbitration.
Correspondingly, each memory pool has two capacities: one is the
current memory pool capacity which can be adjusted under the
memory arbitration during a query execution and the other is the
max capacity which is fixed and configured on the memory pool
creation. The memory arbitrator ensures the memory pool's capacity
is always within the hard configured limit. The memory arbitrator
allocates an initial capacity for the memory pool on its creation and
grows/shrink on-demand then.
If memory arbitrator is not configured, then we always set the memory
pool's capacity to its max capacity.

Also note  both capacity and max capacity apply to the root memory
pool only.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5037

Reviewed By: oerling

Differential Revision: D46139557

Pulled By: xiaoxmeng

fbshipit-source-id: 95c8eddd04ed9aa0433ad14b6e887351a45c3a30