Break HiveConnector.cpp into multiple files (#5175)
Summary:
The HiveConnector.cpp and HiveConnector.h are already very big and contain many classes. This commit breaks it down into several files. The HiveDataSource is in analogous to the existing HiveDataSink files. This is to prepare for the upcoming Iceberg development, such that the files are easier to maintain.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5175

Reviewed By: xiaoxmeng

Differential Revision: D46649354

Pulled By: Yuhta

fbshipit-source-id: e574c4141e733ac6be994fe0b1581c63535ad077