Add dictionary encoded array vector support for FlatMapHelper (#5141)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5141

In order to support the fallback path of batch reuse as well as start validating the correctness in Validation Service ASAP, we will need to support producing regular maps from batch reuse columns.

It is innately trickier to merge dictionary encoded vectors across the different features. Hence, for simplicity, we essentially flatten the dictionary encoding as we produce output map vectors.

Reviewed By: DanielMunozT

Differential Revision: D46426535

fbshipit-source-id: 63f7af8bf5f5af8da2aa89e6fc98d829c09f2c16