Remove nullOutput flag from converter. (#5170)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5170

The flag nullOutput was introduced by https://github.com/facebookincubator/velox/pull/476
as performance optimization.

However its problematic in different ways:
1. It was introduced to remove the cost of the throw->catch->rethrow, however
callers still throw after reading that nullOutput is true, so moving the throw inside the
cast function should be ok.
2. It is only enabled in truncate mode! and only in convertStringToInt conversion, but
not elsewhere!.But the flag spams the converter code while it's not used in almost all
 cast functions.
3. The original diff did not provide evidence of performance.
4. If we want to optimize throw/catch we can pass context and call setError as an
alternative that would avoid the throw if we are inside a try.

Reviewed By: mbasmanova

Differential Revision: D46506557

fbshipit-source-id: b8b8b4d388d3e788a21807bb301406eace8b4501