Fix use-after-free in HTTP callbacks (#19863) (#19865)
Summary:
In the HTTPClient, callbacks are scheduled on an eventBase. HTTPClient is kept alive using a shared_ptr, but it contains a raw pointer to MemoryPool. This MemoryPool may be freed if Task is aborted earlier, but a callback is executed much later.
We see crashes related to this when the batch cluster is under heavy load.

So the fix here is to keep shared_ptr to MemoryPool isntead of a raw pointer

```
== NO RELEASE NOTE ==
```

Pull Request resolved: https://github.com/prestodb/presto/pull/19865

Reviewed By: xiaoxmeng

Differential Revision: D46674355

Pulled By: pranjalssh

fbshipit-source-id: 9b53deb6357ff87b8e1a992f3205d0ce9d79c05c