Fix peeling in FieldReference (#5250)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5250

FieldReference currently uses DecodedVector to implement its own peeling logic and has bugs in it. This diff makes it use PeeledEncoding class for peeling. Note that FieldReference peels encoding off an input RowVector and wrap the encoding to its referenced child, which is not a typical use case of PeeledEncoding. So there are special handling of the child vector being shorter than the top-level vector.

This diff fixes https://github.com/facebookincubator/velox/issues/5248.

Reviewed By: bikramSingh91

Differential Revision: D46469912

fbshipit-source-id: 3e3cef4e4dbc06fe42cd84d90dc978cc4dc6e2c5