Extract row filter from remaining filter (#5258)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5258

Currently if the remaining filter container clauses that are `AND` but
the field is a subfield of struct, or the clause is not `AND` but can be
converted into `AND` (e.g. `NOT (A = 1 OR B = 1)`), Presto optimizer is not able
to identify them.  We do this extraction in Velox so that we don't read more
than necessary.

Reviewed By: oerling

Differential Revision: D46648421

fbshipit-source-id: 94d8b03fbc7f6db2faa416297640e69de1feb0fa