Fix hangs in MultiFragmentTest (#5343)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5343

We've been seeing intermittent hangs in
MultiFragmentTest.raceBetweenTaskTerminateAndTaskNoMoreSplits, e.g.
https://app.circleci.com/pipelines/github/facebookincubator/velox/27164/workflows/b3f6d810-9ea7-4e0c-9513-c187a66cbe5e/jobs/171592

I managed to reproduce locally and found it was stuck waiting on the EventCount
blockTerminate.

Looking at the documentation for the EventCount class, we aren't using it in the
recommended way.  In particular, it's subject to the spurious wake ups and races
that condition variables are, and we need to handle that.
https://github.com/facebook/folly/blob/main/folly/experimental/EventCount.h#L33

Fortunately, there's an await method, which combined with an atomic variable
makes it easy.  I've updated all the EventCount objects in this test and the rest of
the file to use this approach to avoid the race condition we were hitting.

Reviewed By: xiaoxmeng

Differential Revision: D46907187

fbshipit-source-id: d51cc6947f53fef4128ab656f30af884814f0bd8