Add hiveType to HiveColumnHandle (#5335)
Summary:
Add hiveType to HiveColumnHandle which is a velox type parsed
from hive type set in corresponding presto proto. The hiveType
might be different than the dataType in HiveColumnHandle. The
dataType is column type in target write table while hiveType is
column type in source read table. They might not be the same and
table scan need to do the data conversion before sending data to
table writer, and the writer might need to respect the type discrepancy
in data processing such as bucket id calculation. We will add the required
table scan/table writer support later and enforce a check in HiveColumnHandle
to require the two types to be the same.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5335

Reviewed By: mbasmanova

Differential Revision: D46888668

Pulled By: xiaoxmeng

fbshipit-source-id: 8fdae017b423158c68a9bd19f9816a0cd22076d8