Fix assertReadExpected to account for the number of actual rows read in ParquetReaderTestBase (#5320)
Summary:
This PR fixes a small bug in `assertReadExpected` in ParquetReaderTestBase where we
would count the total rows read, i.e. actual rows that match the filter as well as the rows
that were filtered out, thus inflating the final row count. This works fine when the reader
scans all data in the file but breaks once only a subset of rows is requested.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5320

Reviewed By: xiaoxmeng

Differential Revision: D46937330

Pulled By: Yuhta

fbshipit-source-id: 2a60e6b7dead673dc3788ad1656b669d8f723da5