Fix MapFromEntries corner case. (#5344)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5344

For an input of type array(row(int, int)) and values array(null, null, null).
Children of the row can have size 0. map_from_entries throws
error in this case.

But when in a try, those errors are silenced this leads to the construction of
a map vector  that reference elements that do not exist in keys and values.

This diff fix the issue by setting sizes of those indices to 0, this is a fuzzer found issue.
fix the issue: https://github.com/facebookincubator/velox/issues/5342

Reviewed By: kevinwilfong, bikramSingh91

Differential Revision: D46907195

fbshipit-source-id: c0a34992548a22caeab8a3d9c85f866e49c05974