Choose broadcast join when one side is small the other side is unknown
In current cost based join type check, when the estimated size of one
side of the join is small, i.e. within broadcast limit, but the other
side is unknown, we will end up with partitioned join following syntactic
order. This PR adds an option to choose broadcast join with the smaller
side to be build input.

It's controlled by a session parameter which is default to false.