redefine distinctFields and make skipPeeling optimization explicit. (#5354)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5354

distinctFields was defined as
1. the distinct fields of the expression.
2. or empty list if the expression is single referenced and has the same as its parent distinct fields.

(2) was added to optimize peeling (avoiding it) when fields are known to be shared with parents. This diff separate the optimization from the definition of the distinct fields and makes it explicit using skipPeeling() function. With this, users of the library can always trust that distinct fields refer to the expression's distinct fields.

Reviewed By: bikramSingh91

Differential Revision: D46915886

fbshipit-source-id: e2120f89b899730a7addbd083de4d719a7db91b6