More flexible serializer support in remote functions (#5370)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5370

Make remote functions able to specify serializer based on a serde
format in a more flexible manner, without relying on the default serializer
installed in the process. This is needed because different remote functions
might need to communicate with different remote services, using distinct serialization
formats.
Now the serialization format can be defined when registering the function, as
part of the metadata.

Reviewed By: mbasmanova

Differential Revision: D46960082

fbshipit-source-id: 87d876c4d8f577079390a22868576f417207a085