Apply a cost when 0 output variables on TableScan
a query such as:

SELECT 1 from orders, customer;

results in a query plan which reports table scans as having 0 cost.

When calculating query cost, table scans might have 0 output
variables. If this is the case, then the cost of the table scan
node is assigned as 0. This isn't accurate and can result in
sub-optimal plans, especially if there are cross joins.

In the case that a table scan node has no output variables,
we now apply a minimal cost to account for the requirement
of scanning the table.