Enable huge pages for large contiguous allocations (#5413)
Summary:
Hash tables of tens of MB or more benefit from huge pages because of fewer TLB misses. Savings can be ~30% of CPU time.  We enable huge pages for allocations over 2MB (huge page is 2MB). Since this is an advice to the OS that does not change semantics, we do not fail even if the madvise fails.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5413

Reviewed By: xiaoxmeng

Differential Revision: D47033963

Pulled By: oerling

fbshipit-source-id: 690fe1c678847ef3fb29b23e24889943ebac0e3f