Avoid resizing child in FieldReference when parent has encoding layers (#5453)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5453

FieldReference used to resize the child vector through BaseVector::ensureWritable when the child size is smaller than its parent size. This resizing was necessary to avoid errors during when wrapping the encoding layers of the struct over its child where indices in unselected rows can point to indices outside the size of the child. However, this resulted in increased memory usage since the child needed to be copied to resize. This avoids the copy by only operating on the non-null selected rows when peeling and wrapping the child so that those invalid indices are never touched. The final step of adding nulls also occurs after wrapping so that nulls can be directly added over the wrapped layer which would be unique and can be modified in place without the need for a copy.

Reviewed By: kagamiori

Differential Revision: D47110861

fbshipit-source-id: 0daea7d6bf947e0e0600f7e79285327091a85e71