Remove incorrect and confusing code from ValueList (#5475)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5475

> totalBytes_ += stream.size();

This code is incorrect because stream.size() returns total capacity of the stream, not how many bytes have been written.

> auto reserve = std::max<int32_t>(1024, std::min<int64_t>(128, totalBytes_));

This is confusing because the result is always 1024.

The min and max need to be swapped above, but doing this runs in to memory management bug. Hence, just replace with 1024 for now.

Reviewed By: kagamiori

Differential Revision: D47135739

fbshipit-source-id: 025b88fb88d5e0971fb0e33fbd3a309a35d6a9c5