Implement querying and inserts into sample tables
The "samples" table can be accessed similarly to the other table
types in iceberg by using the "$" system table syntax:

presto:tpch> CALL iceberg.system.create_table_sample('tpch', 'customer');

presto:tpch> SELECT * FROM "customer$samples";
 custkey | name | address | nationkey | phone | acctbal | mktsegment | comment
---------+------+---------+-----------+-------+---------+------------+---------
(0 rows)

The sample table always has the same schema as the normal DATA
type table, except during writing and reading the location points the
source and sink to a different location - the location of the sample
table.

One of the tricky parts is that since we're using the system table
syntax to hide the table from the configured catalog, the iceberg
code previously just reads/writes to the iceberg table directly since
a table name like "customer$samples" is parsed into a struct with
something like "{ name -> customer, tableType -> SAMPLES }". The
previous code before this change ignored the table type because
anything other than DATA would return a SystemTable.

SystemTable wasn't a viable implementation because we also needed
write support which SystemTable doesn't provide. So, this
implementation adds a few case statements to check the table type
to see if the sample table needs to be updated as well.

The last bit that needs to be changed after this commit is to support
iceberg's schema evolution. Currently, when querying the table
schema, the schema from the source table is always returned.
However if that schema gets updated, then we need to mirror that
change in the samples table if it exists.

So this implementation resorts to adding